Transfer all update functions to Atualizador
transmitirDados(), gravarDados() and lerTelecomando() are update (threaded) functions, and so were moved from Ajudante() to Atualizador()
